Scrub X-Xsrf-Token and X-Csrf-Token header values in Sentry
