+++ config/sentry.php
@@ -54,6 +54,4 @@ return [
     'traces_sample_rate' => (float) (env('SENTRY_TRACES_SAMPLE_RATE', 0.0)),
 
     'controllers_base_namespace' => env('SENTRY_CONTROLLERS_BASE_NAMESPACE', 'App\\Http\\Controllers'),
-
-    'before_send' => [\App\Sentry\Helpers::class, 'beforeSend'],
 ];
